% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{createdocs}
\alias{createdocs}
\title{Create a new project}
\usage{
createdocs(folder = "./")
}
\arguments{
\item{folder}{the target folder}
}
\description{
create a new project in the a folder
}
\author{
Marcello Gallucci
}
