% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixRd}
\alias{fixRd}
\title{I forgot what it does, but it is useful for extracting help from r packages}
\usage{
fixRd(rd)
}
\arguments{
\item{rd}{an object}
}
\description{
I forgot what it does, but it is useful for extracting help from r packages
}
