% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{foldable_title}
\alias{foldable_title}
\title{create a foldable section with title}
\usage{
foldable_title(title, comment = NULL)
}
\arguments{
\item{title}{a title}

\item{comment}{(optional) comment}
}
\description{
create a foldable section with title
}
