% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_options}
\alias{format_options}
\title{print a list of options and description from yaml file}
\usage{
format_options(com, optnames)
}
\arguments{
\item{com}{the command to look for, as in com.a.yaml}
}
\description{
print a list of options and description from yaml file
}
