% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{include_examples}
\alias{include_examples}
\title{print a html list of pages of category example}
\usage{
include_examples(topic = NULL)
}
\arguments{
\item{topic}{of the file, set in yaml head of the file}
}
\description{
print a html list of pages of category example
}
