% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{include_pages}
\alias{include_pages}
\title{print a html list of pages of category example}
\usage{
include_pages(topic = NULL, category = NULL, text = "", title = "", level = 1)
}
\arguments{
\item{topic}{of the file, set in yaml head of the file}
}
\description{
print a html list of pages of category example
}
