% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{issues}
\alias{issues}
\title{print a html paragraph for issues}
\usage{
issues()
}
\arguments{
\item{topic}{of the file, set in yaml head of the file}
}
\description{
print a html paragraph for issues
}
