% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{link_pages}
\alias{link_pages}
\title{print a link to pages}
\usage{
link_pages(nickname = NULL, topic = NULL, category = NULL, path = ".")
}
\arguments{
\item{nickname}{of the file, set in yaml head of the file}

\item{topic}{of the file, set in yaml head of the file}

\item{category}{of the file, set in yaml head of the file}

\item{path}{path where to find the file, default "."}
}
\description{
print a link to pages
}
