% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_pages}
\alias{list_pages}
\title{print a html list of pages with links}
\usage{
list_pages(nickname = NULL, topic = NULL, category = NULL)
}
\arguments{
\item{nickname}{of the file, set in yaml head of the file}

\item{topic}{of the file, set in yaml head of the file}

\item{category}{of the file, set in yaml head of the file}
}
\description{
print a html list of pages with links
}
