% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{panel_options}
\alias{panel_options}
\title{print a list of options within a panel}
\usage{
panel_options(com, value, labels = FALSE, mode = NULL)
}
\arguments{
\item{com}{the name of the panel to look for, as in .u.yaml}
}
\description{
print a list of options within a panel
}
