% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{jamovi}
\alias{jamovi}
\alias{draft}
\alias{tobecontinued}
\alias{incomplete}
\alias{underconstruction}
\title{shortcuts}
\usage{
`r jamovi`

`r draft`

`r tobecontinued`

`r incomplete`

`r underconstruction`
}
\description{
shortcuts

incomplete warning

under construction warning
}
