% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{include_details}
\alias{include_details}
\title{print a html list of pages of category details}
\usage{
include_details(topic)
}
\arguments{
\item{topic}{of the file, set in yaml head of the file}
}
\description{
print a html list of pages of category details
}
