% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{panel_options}
\alias{panel_options}
\title{print a list of options within a panel}
\usage{
panel_options(com, value)
}
\arguments{
\item{com}{the command to look for, as in com.a.yaml}
}
\description{
print a list of options within a panel
}
