% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{jamovi}
\alias{jamovi}
\alias{draft}
\alias{tobecontinued}
\alias{incomplete}
\title{shortcuts}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
`r jamovi`

`r draft`

`r tobecontinued`

`r incomplete`
}
\description{
shortcuts

incomplete warning
}
\keyword{datasets}
